Add new `pod-info` command.
This command will allow us to gather some pod information filtered by
node.

The pod information gathered is filtered with a go-template. Right now
that template is hard-coded but it would be easy to use an incoming
param to read it from a file name if needed in the future.
Pod information is filtered because we want to avoid to show sensible
information in the must-gather tool.

When, if ever, change the output go-template consider:
1.- this output is going to be parsed at must-gather-pao so any change
should be carefully synced
2.- be sure no sensible information is in the final output (like
environmental variables or command parameters that can contain
passwords)