pfp: recorder: implement maxSize option alongside maxCapacity
Allow to configure the NodeRecorder with maximum capacity to avoid
extreme growth in recorded PFP statuses size. This is yet another
sealing besides maxCapacity that the node recorder should not cross.
The maxSize takes precesdence over maxCapacity, meaning as long as the
current size of the nodeRecorder does not reach the maximum allowed,
`statuses` slice length is maintained and is allowed to reach the
maximum capacity. Once the new item pushed to the nodeRecorder is
expected to make the recorder reach the max allowed size, `statuses`
will witness reductions in older items to allow the new status to fit,
under the constraint that node recorder will at least have one status no
matter its size. So the boundary is important but the recorder goal
should still be on top.

Signed-off-by: Francesco Romani <fromani@redhat.com>