pfp: recorder: support coalescing option
Support new option that if set to true it would configure the recorders
to push PFP statuses only if they are unique in PFP values, otherwise it
just updates the RecordTime of the last recorder status. Note that this
option works only for the last recorded status and it does not affect
other statuses in the stack.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>